♻️ Extract funtion in `commands::install` to fetch both database
